requirements.txt
## The following requirements were added by pip freeze:
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
gspread==3.0.0
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.2
parsel==1.4.0
persian==0.4.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
queuelib==1.5.0
requests==2.19.1
rsa==3.4.2
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
uritemplate==3.0.0
urllib3==1.23
w3lib==1.19.0
zope.interface==4.5.0
